requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scipy
openai
python-dotenv